refactor: refactor the code related to store and styles - add sass dev dependency to the project to write reusable styles code - create a separate file for store base file inside the redux folder instead of having that code in inde.js file - import the store file created newly and use it in index.js file - create color variables in scss and use as constants instead of hard-coding the color variables - add basic scss functions and mixins and reuse those in the UI code
