--- src/App.css
@@ -1,15 +1,0 @@
-.background--green {
-    background-color: green;
-}
-.background--red {
-    background-color: red;
-}
-.background--cyan {
-    background-color: #d3eef1;
-}
-.text--white {
-    color: white;
-}
-.object-fit-cover {
-    object-fit: cover;
-}
