+++ src/assets/styles/_functions.scss
@@ -1,0 +1,3 @@
+@function set-color($color) {
+    @return $color;
+}
