+++ src/assets/styles/_mixins.scss
@@ -1,0 +1,5 @@
+@mixin flex-justify-between {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;    
+}
