+++ src/assets/styles/_variables.scss
@@ -1,0 +1,5 @@
+$white: #ffffff;
+$black: #000000;
+$red: red;
+$green: green;
+$lightBlue: #d3eef1;
