+++ src/index.js
@@ -1,17 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { configureStore } from "@reduxjs/toolkit";
 import { Provider } from "react-redux";
 import "bootstrap/dist/css/bootstrap.min.css";
 import App from "./App";
-import shoppingReducer from "./redux/shoppingSlice";
-
-// configuring the store to be used in the entire application
-const store = configureStore({
-  reducer: {
-    shopping: shoppingReducer,
-  },
-});
+import store from "./redux/store";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
